+++ src/com/kaylerrenslow/armaDialogCreator/data/HeaderConversionException.java
@@ -8,4 +8,8 @@ public class HeaderConversionException extends Exception {
 	public HeaderConversionException(String message) {
 		super(message);
 	}
+
+	public HeaderConversionException(String message, Throwable e) {
+		super(message, e);
+	}
 }